pyproject.toml
[project]
name = "voice"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "duckduckgo-search>=8.1.1",
    "fastapi>=0.116.0",
    "langchain-community>=0.3.27",
    "livekit-agents>=1.1.5",
    "livekit-plugins-google>=1.1.5",
    "livekit-plugins-noise-cancellation>=0.2.5",
    "livekit-plugins-openai>=1.1.5",
    "livekit-plugins-silero>=1.1.5",
    "mem0ai>=0.1.114",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "uvicorn>=0.35.0",
]
